// 코이_이채영/7주차 미션/UMC-mission/src/types/comment.ts
// types/comment.ts

export interface CommentAuthor {
    id: number;
    name: string;
    email: string;
    bio: string | null;
    avatar: string | null;
    createdAt: string;
    updatedAt: string;
}

export interface Comment {
    id: number;
    content: string;
    lpId: number;
    authorId: number;
    createdAt: string;
    updatedAt: string;
    author: CommentAuthor;
}

export interface ResponseCommentListDto {
    status: boolean;
    statusCode: number;
    message: string;
    data: {
        data: Comment[];
        nextCursor: number;
        hasNext: boolean;
    };
}

export interface RequestCommentDto {
    lpId: number;
    content: string;
}

export interface ResponseCommentDto {
    id: number;
    content: string;
    lpId: number;
    authorId: number;
    createdAt: string;
    updatedAt: string;
    author: CommentAuthor;
}